+++ src/Osu.Music.UI/ViewModels/MainViewModel.cs
@@ -11,9 +11,9 @@ using Osu.Music.Services.Interfaces;
 using Osu.Music.Services.IO;
 using Osu.Music.Services.Localization;
 using Osu.Music.Services.UItility;
+using Osu.Music.UI.Behaviors;
 using Osu.Music.UI.Interfaces;
 using Osu.Music.UI.Models;
-using Osu.Music.UI.Utility;
 using Osu.Music.UI.ViewModels.Dialogs;
 using Osu.Music.UI.Views.Dialogs;
 using Osu.Music.UI.Visualization;
